+++ src/main/java/cart/ui/OrderApiController.java
@@ -2,8 +2,8 @@ package cart.ui;
 
 import cart.application.OrderService;
 import cart.domain.member.Member;
-import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
+import cart.dto.order.OrderRequest;
+import cart.dto.order.OrderResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
